requirements.txt
google-tunix[prod]
jax[cpu]
flax
optax
sentencepiece
transformers
torch
numpy
pandas
fastapi
uvicorn
pydantic
pytest
pytest-cov
black
mypy
types-requests
